requirements.txt
streamlit
openai
fpdf2
tavily-python
